Created SingleDomain class
Mostly moved things over from SingleContinuousDomain

Has symbol property, contains method, and simple constructor

added SingleDomain to crv imports (small)
